fix: seperate user data v3
